unified_posadka_bot/requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10.0
pytz>=2023.3
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
Pillow>=10.1.0
numpy>=1.24.3



